requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
alembic
python-dotenv